requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
cachetools==6.2.6
ccxt==4.5.44
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coincurve==21.0.0
cryptography==46.0.5
dateparser==1.3.0
deprecation==2.1.0
distro==1.9.0
frozenlist==1.8.0
fsspec==2026.2.0
groq==1.1.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
postgrest==2.28.3
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python_binance==1.0.36
pytz==2026.1.post1
realtime==2.28.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
